fastapi-app/api/api/api_app.py
from .fastapi_app import create_app
# see https://tillepille.io/posts/otel-fastapi/ since the FastAPIInstrumentor docs are not up to standard
from opentelemetry.instrumentation.fastapi import FastAPIInstrumentor
from opentelemetry import trace
from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import OTLPSpanExporter
from opentelemetry.sdk.trace import TracerProvider
from opentelemetry.sdk.trace.export import BatchSpanProcessor
from opentelemetry.sdk.trace import TracerProvider

app = create_app()

trace.set_tracer_provider(TracerProvider())
tracer = trace.get_tracer(__name__)

otlp_exporter = OTLPSpanExporter()
# we don't want to export every single trace by itself but rather batch them
span_processor = BatchSpanProcessor(otlp_exporter)
otlp_tracer = trace.get_tracer_provider().add_span_processor(span_processor)

FastAPIInstrumentor(app, tracer_provider=otlp_tracer)
