fastapi-app/api/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
requests>=2.31.0
httpx>=0.23.1
gunicorn>=21.2.0
opentelemetry-api>=1.26.0
opentelemetry-sdk>=1.26.0
opentelemetry-exporter-otlp>=1.26.0
opentelemetry-instrumentation-fastapi>=0.47b0
